+++ src/Rector/Contrib/Symfony/FrameworkBundle/AbstractToConstructorInjectionRector.php
@@ -1,0 +1,81 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Rector\Contrib\Symfony\FrameworkBundle;
+
+use PhpParser\Node;
+use PhpParser\Node\Expr\MethodCall;
+use PhpParser\Node\Scalar\String_;
+use Rector\Builder\Class_\ClassPropertyCollector;
+use Rector\Contract\Bridge\ServiceTypeForNameProviderInterface;
+use Rector\Naming\PropertyNaming;
+use Rector\Node\Attribute;
+use Rector\Node\PropertyFetchNodeFactory;
+use Rector\NodeAnalyzer\MethodCallAnalyzer;
+use Rector\Rector\AbstractRector;
+
+abstract class AbstractToConstructorInjectionRector extends AbstractRector
+{
+    /**
+     * @var PropertyNaming
+     */
+    protected $propertyNaming;
+
+    /**
+     * @var ClassPropertyCollector
+     */
+    protected $classPropertyCollector;
+
+    /**
+     * @var PropertyFetchNodeFactory
+     */
+    protected $propertyFetchNodeFactory;
+
+    /**
+     * @var ServiceTypeForNameProviderInterface
+     */
+    protected $serviceTypeForNameProvider;
+
+    /**
+     * @var MethodCallAnalyzer
+     */
+    protected $methodCallAnalyzer;
+
+    public function __construct(
+        PropertyNaming $propertyNaming,
+        ClassPropertyCollector $classPropertyCollector,
+        PropertyFetchNodeFactory $propertyFetchNodeFactory,
+        ServiceTypeForNameProviderInterface $serviceTypeForNameProvider,
+        MethodCallAnalyzer $methodCallAnalyzer
+    ) {
+        $this->propertyNaming = $propertyNaming;
+        $this->classPropertyCollector = $classPropertyCollector;
+        $this->propertyFetchNodeFactory = $propertyFetchNodeFactory;
+        $this->serviceTypeForNameProvider = $serviceTypeForNameProvider;
+        $this->methodCallAnalyzer = $methodCallAnalyzer;
+    }
+
+    /**
+     * @param MethodCall $methodCallNode
+     */
+    public function refactor(Node $methodCallNode): ?Node
+    {
+        /** @var String_ $stringArgument */
+        $stringArgument = $methodCallNode->args[0]->value;
+
+        $serviceName = $stringArgument->value;
+        $serviceType = $this->serviceTypeForNameProvider->provideTypeForName($serviceName);
+        if ($serviceType === null) {
+            return null;
+        }
+
+        $propertyName = $this->propertyNaming->typeToName($serviceType);
+
+        $this->classPropertyCollector->addPropertyForClass(
+            (string) $methodCallNode->getAttribute(Attribute::CLASS_NAME),
+            [$serviceType],
+            $propertyName
+        );
+
+        return $this->propertyFetchNodeFactory->createLocalWithPropertyName($propertyName);
+    }
+}
