+++ src/Rector/Contrib/Symfony/FrameworkBundle/ContainerGetToConstructorInjectionRector.php
@@ -3,15 +3,7 @@
 namespace Rector\Rector\Contrib\Symfony\FrameworkBundle;
 
 use PhpParser\Node;
-use PhpParser\Node\Expr\MethodCall;
-use PhpParser\Node\Scalar\String_;
-use Rector\Builder\Class_\ClassPropertyCollector;
-use Rector\Contract\Bridge\ServiceTypeForNameProviderInterface;
-use Rector\Naming\PropertyNaming;
 use Rector\Node\Attribute;
-use Rector\Node\PropertyFetchNodeFactory;
-use Rector\NodeAnalyzer\MethodCallAnalyzer;
-use Rector\Rector\AbstractRector;
 
 /**
  * Ref: https://github.com/symfony/symfony/blob/master/UPGRADE-4.0.md#console
@@ -36,33 +28,8 @@ use Rector\Rector\AbstractRector;
  *      $this->someService
  * }
  */
-final class ContainerGetToConstructorInjectionRector extends AbstractRector
+final class ContainerGetToConstructorInjectionRector extends AbstractToConstructorInjectionRector
 {
-    /**
-     * @var ClassPropertyCollector
-     */
-    private $classPropertyCollector;
-
-    /**
-     * @var PropertyNaming
-     */
-    private $propertyNaming;
-
-    /**
-     * @var PropertyFetchNodeFactory
-     */
-    private $propertyFetchNodeFactory;
-
-    /**
-     * @var ServiceTypeForNameProviderInterface
-     */
-    private $serviceTypeForNameProvider;
-
-    /**
-     * @var MethodCallAnalyzer
-     */
-    private $methodCallAnalyzer;
-
     /**
      * @var string[]
      */
@@ -70,20 +37,6 @@ final class ContainerGetToConstructorInjectionRector extends AbstractRector
         'Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand',
         'Symfony\Bundle\FrameworkBundle\Controller\Controller',
     ];
-
-    public function __construct(
-        ClassPropertyCollector $classPropertyCollector,
-        PropertyNaming $propertyNaming,
-        PropertyFetchNodeFactory $propertyFetchNodeFactory,
-        ServiceTypeForNameProviderInterface $serviceTypeForNameProvider,
-        MethodCallAnalyzer $methodCallAnalyzer
-    ) {
-        $this->classPropertyCollector = $classPropertyCollector;
-        $this->propertyNaming = $propertyNaming;
-        $this->propertyFetchNodeFactory = $propertyFetchNodeFactory;
-        $this->serviceTypeForNameProvider = $serviceTypeForNameProvider;
-        $this->methodCallAnalyzer = $methodCallAnalyzer;
-    }
 
     public function isCandidate(Node $node): bool
     {
@@ -98,30 +51,5 @@ final class ContainerGetToConstructorInjectionRector extends AbstractRector
         $parentClassName = $node->getAttribute(Attribute::PARENT_CLASS_NAME);
 
         return in_array($parentClassName, $this->containerAwareParentTypes, true);
-    }
-
-    /**
-     * @param MethodCall $methodCallNode
-     */
-    public function refactor(Node $methodCallNode): ?Node
-    {
-        /** @var String_ $serviceNameArgument */
-        $serviceNameArgument = $methodCallNode->args[0]->value;
-
-        $serviceName = $serviceNameArgument->value;
-        $serviceType = $this->serviceTypeForNameProvider->provideTypeForName($serviceName);
-        if ($serviceType === null) {
-            return null;
-        }
-
-        $propertyName = $this->propertyNaming->typeToName($serviceType);
-
-        $this->classPropertyCollector->addPropertyForClass(
-            (string) $methodCallNode->getAttribute(Attribute::CLASS_NAME),
-            [$serviceType],
-            $propertyName
-        );
-
-        return $this->propertyFetchNodeFactory->createLocalWithPropertyName($propertyName);
     }
 }
