+++ src/Rector/Contrib/Symfony/FrameworkBundle/GetToConstructorInjectionRector.php
@@ -4,14 +4,6 @@ namespace Rector\Rector\Contrib\Symfony\FrameworkBundle;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\MethodCall;
-use PhpParser\Node\Scalar\String_;
-use Rector\Builder\Class_\ClassPropertyCollector;
-use Rector\Contract\Bridge\ServiceTypeForNameProviderInterface;
-use Rector\Naming\PropertyNaming;
-use Rector\Node\Attribute;
-use Rector\Node\PropertyFetchNodeFactory;
-use Rector\NodeAnalyzer\MethodCallAnalyzer;
-use Rector\Rector\AbstractRector;
 
 /**
  * Before:
@@ -20,47 +12,8 @@ use Rector\Rector\AbstractRector;
  * After:
  * - $this->someService # where "someService" is type of the service
  */
-final class GetToConstructorInjectionRector extends AbstractRector
+final class GetToConstructorInjectionRector extends AbstractToConstructorInjectionRector
 {
-    /**
-     * @var PropertyNaming
-     */
-    private $propertyNaming;
-
-    /**
-     * @var ClassPropertyCollector
-     */
-    private $classPropertyCollector;
-
-    /**
-     * @var PropertyFetchNodeFactory
-     */
-    private $propertyFetchNodeFactory;
-
-    /**
-     * @var ServiceTypeForNameProviderInterface
-     */
-    private $serviceTypeForNameProvider;
-
-    /**
-     * @var MethodCallAnalyzer
-     */
-    private $methodCallAnalyzer;
-
-    public function __construct(
-        PropertyNaming $propertyNaming,
-        ClassPropertyCollector $classPropertyCollector,
-        PropertyFetchNodeFactory $propertyFetchNodeFactory,
-        ServiceTypeForNameProviderInterface $serviceTypeForNameProvider,
-        MethodCallAnalyzer $methodCallAnalyzer
-    ) {
-        $this->propertyNaming = $propertyNaming;
-        $this->classPropertyCollector = $classPropertyCollector;
-        $this->propertyFetchNodeFactory = $propertyFetchNodeFactory;
-        $this->serviceTypeForNameProvider = $serviceTypeForNameProvider;
-        $this->methodCallAnalyzer = $methodCallAnalyzer;
-    }
-
     public function isCandidate(Node $node): bool
     {
         if (! $node instanceof MethodCall) {
@@ -72,30 +25,5 @@ final class GetToConstructorInjectionRector extends AbstractRector
             'Symfony\Bundle\FrameworkBundle\Controller\Controller',
             'get'
         );
-    }
-
-    /**
-     * @param MethodCall $methodCallNode
-     */
-    public function refactor(Node $methodCallNode): ?Node
-    {
-        /** @var String_ $stringArgument */
-        $stringArgument = $methodCallNode->args[0]->value;
-
-        $serviceName = $stringArgument->value;
-        $serviceType = $this->serviceTypeForNameProvider->provideTypeForName($serviceName);
-        if ($serviceType === null) {
-            return null;
-        }
-
-        $propertyName = $this->propertyNaming->typeToName($serviceType);
-
-        $this->classPropertyCollector->addPropertyForClass(
-            (string) $methodCallNode->getAttribute(Attribute::CLASS_NAME),
-            [$serviceType],
-            $propertyName
-        );
-
-        return $this->propertyFetchNodeFactory->createLocalWithPropertyName($propertyName);
     }
 }
